pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chat-api"
version = "1.0.0"
description = "A simple LLM chat service with Pythonic design"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Chat API Team"}]
keywords = ["api", "llm", "chat", "fastapi", "python"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "databases[sqlite,postgresql]",
    "redis[hiredis]",
    "litellm",
    "slowapi",
    "tenacity",
    "pydantic-settings",
    "sqlalchemy",
    "loguru",
    "orjson",
    "boto3",  # For DynamoDB support
    "aioboto3",  # For async DynamoDB operations
    "mangum",  # For Lambda deployment
    "prometheus-client",  # For litellm Prometheus integration
    "backoff",  # Required by litellm for retries
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
    "ruff>=0.8.0",
    "mypy>=1.0",
    "bandit[toml]>=1.7",
    "types-redis",
]

test = [
    "pytest>=8.0",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
]

lambda = [
    "mangum",
]

[project.scripts]
chat-api = "chat_api.__main__:main"

[project.urls]
Homepage = "https://github.com/your-org/chat-api"
Repository = "https://github.com/your-org/chat-api.git"
Issues = "https://github.com/your-org/chat-api/issues"

[tool.ruff]
target-version = "py311"
line-length = 100
fix = true

[tool.ruff.lint]
select = [
    # Core error checking
    "E4", "E7", "E9", "F",
    # Bug prevention
    "B", "BLE", "TRY",
    # Performance & Security
    "S", "G", "PERF",
    # Code quality
    "SIM", "RET", "RUF", "PTH",
    # FastAPI/Modern Python
    "ASYNC", "I", "UP", "TCH",
]

ignore = [
    "E501",    # Line length (formatter handles)
    "S101",    # Assert in tests
    "S104",    # Bind 0.0.0.0 (Docker)
    "TRY003",  # Long exception messages
    "B008",    # FastAPI Depends()
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S", "ARG", "PLR2004"]
"*/lambda_handler.py" = ["G004"]

[tool.ruff.lint.isort]
known-first-party = ["chat_api"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
check_untyped_defs = true
no_implicit_optional = true
warn_no_return = true
warn_unreachable = true
disallow_untyped_defs = false
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
addopts = [
    "-ra",
    "--strict-markers",
    "--tb=short",
    "-q",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: integration tests requiring external services",
    "unit: unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning:pydantic.*",
    "ignore::DeprecationWarning:httpx.*",
]

[tool.coverage.run]
source = ["chat_api"]
branch = true
omit = [
    "*/tests/*",
    "*/__main__.py",
    "*/lambda_handler.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "except ImportError:",
]
precision = 2
show_missing = true
skip_covered = false
fail_under = 75

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]

[tool.hatch.build.targets.wheel]
packages = ["chat_api"]

[tool.hatch.build.targets.sdist]
include = ["/chat_api", "/tests", "/README.md", "/LICENSE"]

[dependency-groups]
dev = [
    "bandit[toml]>=1.8.6",
    "httpx>=0.28.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "types-redis>=4.6.0.20241004",
]
